refactor: consolidate DOM types and remove dead code
- Move DOM types from src/types/dom.ts into main src/types.ts
- Remove deprecated src/utils/validation.ts (replaced by commands/shared/validation.ts)
- Remove unused filterOption from commonOptions.ts
- Remove truncateText re-export wrapper from remoteObject.ts
- Fix TSDoc syntax in test file (arrow to text)
