+++ src/session/QueryCacheManager.ts
@@ -30,7 +30,7 @@ import { join } from 'path';
 
 import { getErrorMessage } from '@/connection/errors.js';
 import { getSessionDir } from '@/session/paths.js';
-import type { DomQueryResult } from '@/types/dom.js';
+import type { DomQueryResult } from '@/types.js';
 import { createLogger } from '@/ui/logging/index.js';
 
 const log = createLogger('session');
